fix: Update type declarations to use float instead of int
Update the type declarations in the `proportional` methods of `Exclusive.php` and `Inclusive.php` files to use float instead of int for the `$percentage` parameter. This ensures more accurate calculations for tax percentages.

+++ src/Adjustments/Taxation/Exclusive.php
@@ -14,7 +14,7 @@ use Vaened\PriceEngine\Adjustments\AdjustmentType;
 
 final class Exclusive extends TaxScheme
 {
-    public static function proportional(int $percentage, BackedEnum|UnitEnum|string $code): self
+    public static function proportional(float $percentage, BackedEnum|UnitEnum|string $code): self
     {
         return new self(AdjustmentType::Percentage, $percentage, AdjustmentMode::PerUnit, $code);
     }
